added header to medium posts. udpate deps
