+++ package.json
@@ -49,7 +49,7 @@
     "postcss-simple-vars": "^7.0.1",
     "remark-frontmatter": "^5.0.0",
     "remark-mdx-frontmatter": "^5.0.0",
-    "typescript": "^5.5.3",
+    "typescript": "~5.4.0",
     "vite": "^5.3.3",
     "vite-tsconfig-paths": "^4.3.2"
   },
